also utilize float32 where default is float64

+++ npbench/benchmarks/cavity_flow/cavity_flow.py
@@ -7,7 +7,7 @@ def initialize(ny, nx):
     u = np.zeros((ny, nx), dtype=np.float32)
     v = np.zeros((ny, nx), dtype=np.float32)
     p = np.zeros((ny, nx), dtype=np.float32)
-    dx = 2 / (nx - 1)
-    dy = 2 / (ny - 1)
-    dt = .1 / ((nx - 1) * (ny - 1))
+    dx = np.float32(2 / (nx - 1))
+    dy = np.float32(2 / (ny - 1))
+    dt = np.float32(.1 / ((nx - 1) * (ny - 1)))
     return u, v, p, dx, dy, dt
