+++ npbench/benchmarks/weather_stencils/vadv/vadv.py
@@ -7,13 +7,13 @@ def initialize(I, J, K):
     from numpy.random import default_rng
     rng = default_rng(42)
 
-    dtr_stage = 3. / 20.
+    dtr_stage = np.float32(3. / 20.)
 
     # Define arrays
-    utens_stage = rng.random((I, J, K))
-    u_stage = rng.random((I, J, K))
-    wcon = rng.random((I + 1, J, K))
-    u_pos = rng.random((I, J, K))
-    utens = rng.random((I, J, K))
+    utens_stage = rng.random((I, J, K), dtype=np.float32)
+    u_stage = rng.random((I, J, K), dtype=np.float32)
+    wcon = rng.random((I + 1, J, K), dtype=np.float32)
+    u_pos = rng.random((I, J, K), dtype=np.float32)
+    utens = rng.random((I, J, K), dtype=np.float32)
 
     return dtr_stage, utens_stage, u_stage, wcon, u_pos, utens
